+++ CHANGELOG.md
@@ -103,3 +103,8 @@
 - Modul Attendance Mingguan (feedback MJ)
     - Notifikasi Duplikasi data
     - Add component overlay loading
+
+# 0.0.1 (Update 07/02/2023)
+- Modul Attendance Mingguan (feedback MJ)
+    - Filter Attendance
+    - Tambah, Hapus, dan Edit kehadiran by filter